requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
starlette
supabase
PyJWT
python-dotenv
Pillow
